scripts/deforestation_analysis.py
#!/usr/bin/env python3
"""
Deforestation Analysis Script
----------------------------
Analyzes historical deforestation patterns using Hansen Global Forest Change dataset.
"""

import os
import numpy as np
import pandas as pd
import geopandas as gpd
import rasterio
from rasterio.mask import mask
import earthengine as ee
from datetime import datetime

class DeforestationAnalyzer:
    def __init__(self, aoi_path, output_dir):
        """
        Initialize the analyzer with Area of Interest (AOI) and output directory.
        
        Args:
            aoi_path (str): Path to AOI shapefile
            output_dir (str): Directory for output files
        """
        self.aoi_path = aoi_path
        self.output_dir = output_dir
        self.aoi = None
        self.hansen_data = None
        
    def load_aoi(self):
        """Load and validate the Area of Interest shapefile."""
        try:
            self.aoi = gpd.read_file(self.aoi_path)
            print(f"Loaded AOI with {len(self.aoi)} features")
        except Exception as e:
            print(f"Error loading AOI: {e}")
            raise
            
    def initialize_earth_engine(self):
        """Initialize Google Earth Engine connection."""
        try:
            ee.Initialize()
            print("Successfully initialized Earth Engine")
        except Exception as e:
            print(f"Error initializing Earth Engine: {e}")
            raise
            
    def get_hansen_data(self, year_start=2001, year_end=None):
        """
        Fetch Hansen Global Forest Change data for the specified period.
        
        Args:
            year_start (int): Start year for analysis
            year_end (int): End year for analysis (defaults to current year)
        """
        if year_end is None:
            year_end = datetime.now().year - 1
            
        try:
            dataset = ee.ImageCollection('UMD/hansen/global_forest_change')
            self.hansen_data = dataset.filterDate(f'{year_start}-01-01', f'{year_end}-12-31')
            print(f"Retrieved Hansen data for {year_start}-{year_end}")
        except Exception as e:
            print(f"Error fetching Hansen data: {e}")
            raise
            
    def calculate_forest_loss(self):
        """Calculate annual forest loss within the AOI."""
        # Implementation details to be added
        pass
        
    def generate_statistics(self):
        """Generate summary statistics of forest loss."""
        # Implementation details to be added
        pass
        
    def export_results(self):
        """Export analysis results to files."""
        # Implementation details to be added
        pass

def main():
    """Main execution function."""
    # Example usage
    analyzer = DeforestationAnalyzer(
        aoi_path="path/to/aoi.shp",
        output_dir="path/to/output"
    )
    
    # Run analysis
    analyzer.load_aoi()
    analyzer.initialize_earth_engine()
    analyzer.get_hansen_data()
    analyzer.calculate_forest_loss()
    analyzer.generate_statistics()
    analyzer.export_results()

if __name__ == "__main__":
    main()