requirements.txt
numpy>=1.24.0
pandas>=2.0.0
geopandas>=0.13.0
rasterio>=1.3.6
earthengine-api>=0.1.350
folium>=0.14.0
jupyter>=1.0.0
matplotlib>=3.7.1
seaborn>=0.12.2
scikit-learn>=1.2.2
xarray>=2023.1.0
rioxarray>=0.14.1
google-cloud-storage>=2.8.0
shapely>=2.0.1
fiona>=1.9.3
pyproj>=3.5.0
rasterstats>=0.19.0